Adding names listener.
The server sends names as a 353 (with a 366 to mark the end of the list, if that matters).
The server will send multiple 353s for a single names command if there are enough people on the channel.
The parser doesn't get a channel from that, so I check the params to find the appropriate channel.
The text is a list of names, which may need to have some characters stripped (like @ for ops).

I don't care right now about duplicate entries in the list, since I invision using the list just
to see whether or not someone is on the channel. If it's a problem, the array can get some uniq
love, or it could just become a Set.