check if internal and external encoding are the same
If setting both internal_encoding and external_encoding to the same
encoding, Ruby will raise an exception because it can't convert
between the two. But apparantly, some users have such a setup, so we
explicitly check for that and don't set an internal encoding for the
socket if it'd equal the external encoding.

Closes gh-10