Remove User#hash method
When we initially wrote User#hash, we did not realize that a user's
nickname might change during runtime (which it obviously does), and
that it'd break hash indexing and possibly other stuff.

Nowadays, the solution to fixing this problem is to discourage people
from using User.new. Instead we are relying on a Manager that hands
out one and the same User object.