use a valid server name in WHOIS
Usually, it doesn't matter what you specify as the server in a WHOIS.
But in one case, freenode was sending us a "no such server" error that
didn't include the server name (then the nick) and thus break synced
attributes.

By specifying a real server, we'll always get a "no such nick" instead
of a "no such server".