Allow unused parameters in MODE for Slack
Slack sends weird mode strings like "+v user : active", which leaves us
with an additional parameter ("active"). Normally, we raise exceptions
on invalid mode strings, to discover bugs in both our implementation as
well as that of IRCds.

As I don't expect Slack to fix this issue, and because there'll likely
be more issues in the future, we add special support for Slack and
ignore this error.