Used sysread and smaller buffer
If i use sysread / syswrite and open the file with "rb" in the plugin. I could transfer the data 10 out of 10 times.
I am not sure where the problem is. If i do tcpdump on both ends, it works with the old version also.
As i am not a ruby coder, i am unable to say if this is a "solution"